resnet with 200 classes

+++ p81/training/training.py
@@ -12,7 +12,7 @@ def train(model, device, train_loader, loss_function, optimizer, epoch,lambda_l1
         for batch_idx, (data, target) in enumerate(tepoch):
             loss = 0
             tepoch.set_description(f"Train Epoch {epoch}")
-            data, target = data.to(device), target.to(device)
+            data, target = data.float().to(device), target.to(device)
             optimizer.zero_grad()
 #            print("sending data to model")
             output = model(data)
